Improved accuracy of frame and render times in the editor

+++ Runtime/Core/Stopwatch.h
@@ -30,9 +30,9 @@ namespace Directus
 		~Stopwatch();
 
 		void Start();
-		double Stop();
+		float Stop();
 
 	private:
-		double m_milliseconds;
+		float m_milliseconds;
 	};
 }
